requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
sentence-transformers==3.0.1
chromadb==0.5.5
python-dotenv==1.0.1
rank-bm25==0.2.2
numpy==1.26.4
pydantic==2.9.2
torch==2.4.0  # Optional for GPU acceleration in sentence-transformers